+++ dms-application/src/main/kotlin/team/aliens/dms/common/extension/LocalDateExtension.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.common
+package team.aliens.dms.common.extension
 
 import java.time.LocalDate
