+++ dms-application/src/main/kotlin/team/aliens/dms/domain/meal/usecase/QueryMealsUseCase.kt
@@ -1,7 +1,7 @@
 package team.aliens.dms.domain.meal.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
-import team.aliens.dms.common.iterator
+import team.aliens.dms.common.extension.iterator
 import team.aliens.dms.domain.meal.dto.QueryMealsResponse
 import team.aliens.dms.domain.meal.dto.QueryMealsResponse.MealDetails
 import team.aliens.dms.domain.meal.spi.MealQueryStudentPort
